+++ testdata/de/dm/intellij/liferay/language/osgi/MetaConfigurationInvalidIdInspectionTest/de/dm/configuration/MyConfiguration.java
@@ -1,0 +1,7 @@
+package de.dm.configuration;
+
+import aQute.bnd.annotation.metatype.Meta;
+
+@Meta.OCD(id = <warning descr="ID does not match full qualified class name">"de.dm.configuration.MyInvalidConfiguration"</warning>)
+public interface MyConfiguration {
+}