+++ testdata/de/dm/intellij/liferay/language/osgi/MetaConfigurationInvalidIdInspectionTest/de/dm/configuration/MyValidConfiguration.java
@@ -1,0 +1,7 @@
+package de.dm.configuration;
+
+import aQute.bnd.annotation.metatype.Meta;
+
+@Meta.OCD(id = "de.dm.configuration.MyValidConfiguration")
+public interface MyValidConfiguration {
+}